+++ tooling/hcpctl/internal/e2e/e2e_test.go
@@ -12,18 +12,212 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build E2Etests
+//go:build ToolingE2Etests
 
 package e2e
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"os"
 	"os/exec"
+	"path/filepath"
 	"strings"
 	"testing"
+	"time"
+
+	"gopkg.in/yaml.v2"
+
+	"github.com/Azure/ARO-HCP/tooling/hcpctl/pkg/release"
 )
+
+const (
+	// Test configuration
+	DefaultTimeout = 1 * time.Minute
+	HCPTimeout     = 1 * time.Minute
+
+	// Binary path
+	HCPCTLBinary = "../../hcpctl"
+)
+
+var (
+	DefaultMCCluster = os.Getenv("E2E_MC_CLUSTER")
+	DefaultSCCluster = os.Getenv("E2E_SC_CLUSTER")
+)
+
+// TestCluster represents cluster information for testing
+type TestCluster struct {
+	Name string
+	Type string // "mc", "sc", "hcp"
+}
+
+// AuthOutput represents the structure of kubectl auth whoami -o yaml output
+type AuthOutput struct {
+	APIVersion string `yaml:"apiVersion"`
+	Kind       string `yaml:"kind"`
+	Status     struct {
+		UserInfo struct {
+			Username string              `yaml:"username"`
+			Groups   []string            `yaml:"groups"`
+			Extra    map[string][]string `yaml:"extra,omitempty"`
+		} `yaml:"userInfo"`
+	} `yaml:"status"`
+}
+
+// ClusterListWrapper represents the wrapper structure for cluster list outputs
+type ClusterListWrapper struct {
+	Items []interface{} `yaml:"items"`
+}
+
+// SCCluster represents SC cluster information from sc list output
+type SCCluster struct {
+	Name           string `yaml:"name" json:"name"`
+	Location       string `yaml:"location" json:"location"`
+	ResourceGroup  string `yaml:"resourcegroup" json:"resourcegroup"`
+	SubscriptionID string `yaml:"subscriptionid" json:"subscriptionid"`
+	State          string `yaml:"state" json:"state"`
+}
+
+// MCCluster represents MC cluster information from mc list output
+type MCCluster struct {
+	Name           string `yaml:"name" json:"name"`
+	Location       string `yaml:"location" json:"location"`
+	ResourceGroup  string `yaml:"resourcegroup" json:"resourcegroup"`
+	SubscriptionID string `yaml:"subscription" json:"subscription"`
+	State          string `yaml:"state" json:"state"`
+}
+
+// HCPCluster represents HCP cluster information from hcp list output
+type HCPCluster struct {
+	ID             string `yaml:"id"`
+	Name           string `yaml:"name"`
+	SubscriptionID string `yaml:"subscriptionId"`
+}
+
+// execCommand executes a command with timeout and returns output
+func execCommand(ctx context.Context, name string, args ...string) ([]byte, error) {
+	cmd := exec.CommandContext(ctx, name, args...)
+	cmd.Env = os.Environ()
+	return cmd.Output()
+}
+
+// execCommandWithKubeconfig executes a command with a specific kubeconfig
+func execCommandWithKubeconfig(ctx context.Context, kubeconfig, name string, args ...string) ([]byte, error) {
+	cmd := exec.CommandContext(ctx, name, args...)
+	cmd.Env = append(os.Environ(), fmt.Sprintf("KUBECONFIG=%s", kubeconfig))
+	return cmd.Output()
+}
+
+// execCommandWithKubeconfigAndStderr executes a command with a specific kubeconfig and captures both stdout and stderr
+func execCommandWithKubeconfigAndStderr(ctx context.Context, kubeconfig, name string, args ...string) ([]byte, []byte, error) {
+	cmd := exec.CommandContext(ctx, name, args...)
+	cmd.Env = append(os.Environ(), fmt.Sprintf("KUBECONFIG=%s", kubeconfig))
+	stdout, err := cmd.Output()
+	var stderr []byte
+	if exitError, ok := err.(*exec.ExitError); ok {
+		stderr = exitError.Stderr
+	}
+	return stdout, stderr, err
+}
+
+// parseAuthOutput parses kubectl auth whoami YAML output
+func parseAuthOutput(data []byte) (*AuthOutput, error) {
+	var auth AuthOutput
+	if err := yaml.Unmarshal(data, &auth); err != nil {
+		return nil, fmt.Errorf("failed to parse auth output: %w", err)
+	}
+	return &auth, nil
+}
+
+// parseSCListYAML parses sc list YAML output
+func parseSCListYAML(data []byte) ([]SCCluster, error) {
+	var wrapper struct {
+		Items []SCCluster `yaml:"items"`
+	}
+	if err := yaml.Unmarshal(data, &wrapper); err != nil {
+		return nil, fmt.Errorf("failed to parse SC list YAML output: %w", err)
+	}
+	return wrapper.Items, nil
+}
+
+// parseSCListJSON parses sc list JSON output
+func parseSCListJSON(data []byte) ([]SCCluster, error) {
+	var wrapper struct {
+		Items []SCCluster `json:"items"`
+	}
+	if err := json.Unmarshal(data, &wrapper); err != nil {
+		return nil, fmt.Errorf("failed to parse SC list JSON output: %w", err)
+	}
+	return wrapper.Items, nil
+}
+
+// parseSCList parses sc list output (legacy, defaults to YAML)
+func parseSCList(data []byte) ([]SCCluster, error) {
+	return parseSCListYAML(data)
+}
+
+// parseMCListYAML parses mc list YAML output
+func parseMCListYAML(data []byte) ([]MCCluster, error) {
+	var wrapper struct {
+		Items []MCCluster `yaml:"items"`
+	}
+	if err := yaml.Unmarshal(data, &wrapper); err != nil {
+		return nil, fmt.Errorf("failed to parse MC list YAML output: %w", err)
+	}
+	return wrapper.Items, nil
+}
+
+// parseMCListJSON parses mc list JSON output
+func parseMCListJSON(data []byte) ([]MCCluster, error) {
+	var wrapper struct {
+		Items []MCCluster `json:"items"`
+	}
+	if err := json.Unmarshal(data, &wrapper); err != nil {
+		return nil, fmt.Errorf("failed to parse MC list JSON output: %w", err)
+	}
+	return wrapper.Items, nil
+}
+
+// parseHCPList parses hcp list YAML output
+func parseHCPList(data []byte) ([]HCPCluster, error) {
+	var wrapper struct {
+		Items []HCPCluster `yaml:"items"`
+	}
+	if err := yaml.Unmarshal(data, &wrapper); err != nil {
+		return nil, fmt.Errorf("failed to parse HCP list output: %w", err)
+	}
+	return wrapper.Items, nil
+}
+
+// hasGroup checks if the user has any of the specified groups
+func hasGroup(auth *AuthOutput, groups ...string) bool {
+	userGroups := auth.Status.UserInfo.Groups
+	for _, userGroup := range userGroups {
+		for _, requiredGroup := range groups {
+			if strings.Contains(userGroup, requiredGroup) {
+				return true
+			}
+		}
+	}
+	return false
+}
+
+// parseReleaseStatusYAML parses hcpctl release status YAML output
+func parseReleaseStatusYAML(data []byte) (*release.ClusterComponentRelease, error) {
+	var clusterRelease release.ClusterComponentRelease
+	if err := yaml.Unmarshal(data, &clusterRelease); err != nil {
+		return nil, fmt.Errorf("failed to parse release status YAML output: %w", err)
+	}
+	return &clusterRelease, nil
+}
+
+// createTempKubeconfig creates a temporary kubeconfig file path
+func createTempKubeconfig(t *testing.T, prefix string) string {
+	t.Helper()
+	tempDir := t.TempDir()
+	return filepath.Join(tempDir, fmt.Sprintf("%s-kubeconfig.yaml", prefix))
+}
 
 func TestMain(m *testing.M) {
 	// Check if binary exists, if not, try to build it
@@ -530,7 +724,7 @@ func TestHCPBreakglassRole(t *testing.T) {
 		t.Skip("No HCP clusters available for testing")
 	}
 
-	hcpCluster := hcpClusters[0]
+	hcpCluster := hcpClusters[len(hcpClusters)-1]
 	t.Logf("Using HCP cluster: %s (name: %s)", hcpCluster.ID, hcpCluster.Name)
 
 	tests := []struct {
@@ -675,6 +869,111 @@ func TestHCPList(t *testing.T) {
 			t.Logf("HCP list successful: found %d clusters", len(hcpClusters))
 			for i, cluster := range hcpClusters {
 				t.Logf("  [%d] ID: %s, Name: %s, Subscription: %s", i+1, cluster.ID, cluster.Name, cluster.SubscriptionID)
+			}
+		})
+	}
+}
+
+func TestReleaseStatus(t *testing.T) {
+	tests := []struct {
+		name         string
+		outputFormat string
+		releaseName  string
+		namespace    string
+	}{
+		{
+			name:         "release status in SC context",
+			outputFormat: "yaml",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+			ctx, cancel := context.WithTimeout(context.Background(), DefaultTimeout)
+			defer cancel()
+
+			// Create a temporary kubeconfig for this test
+			kubeconfigPath := createTempKubeconfig(t, "sc-release-status")
+
+			// Get SC access via breakglass
+			_, err := execCommand(ctx, HCPCTLBinary, "sc", "breakglass", DefaultSCCluster, "--no-shell", "--output", kubeconfigPath)
+			if err != nil {
+				t.Fatalf("sc breakglass failed: %v", err)
+			}
+
+			// Verify kubeconfig was created
+			if _, err := os.Stat(kubeconfigPath); os.IsNotExist(err) {
+				t.Fatal("kubeconfig file was not created")
+			}
+
+			// Build release status command args
+			args := []string{"release", "status", "--kubeconfig", kubeconfigPath, "-o", tt.outputFormat}
+			if tt.releaseName != "" {
+				args = append(args, "--release", tt.releaseName)
+			}
+			if tt.namespace != "" {
+				args = append(args, "--namespace", tt.namespace)
+			}
+
+			// Execute release status command
+			output, err := execCommand(ctx, HCPCTLBinary, args...)
+
+			if err != nil {
+				t.Fatalf("release status failed: %v", err)
+			}
+
+			if len(output) == 0 {
+				t.Fatal("release status returned empty output")
+			}
+
+			// Parse
+			clusterRelease, err := parseReleaseStatusYAML(output)
+			if err != nil {
+				t.Fatalf("failed to parse release status YAML: %v", err)
+			}
+
+			// Validate the structure
+			if clusterRelease.APIVersion != "service-status.hcm.openshift.io/v1" {
+				t.Errorf("unexpected APIVersion: got %s, want service-status.hcm.openshift.io/v1", clusterRelease.APIVersion)
+			}
+			if clusterRelease.Kind != "ClusterComponentRelease" {
+				t.Errorf("unexpected Kind: got %s, want ClusterComponentRelease", clusterRelease.Kind)
+			}
+			if clusterRelease.Metadata.Name == "" {
+				t.Error("metadata.name is empty")
+			}
+			if clusterRelease.Metadata.CreationTimestamp.IsZero() {
+				t.Error("metadata.creationTimestamp is zero")
+			}
+
+			// Validate each component has expected structure
+			for i, component := range clusterRelease.Components {
+				if component.APIVersion != "service-status.hcm.openshift.io/v1" {
+					t.Errorf("component %d has unexpected APIVersion: got %s, want service-status.hcm.openshift.io/v1", i, component.APIVersion)
+				}
+				if component.Kind != "ComponentRelease" {
+					t.Errorf("component %d has unexpected Kind: got %s, want ComponentRelease", i, component.Kind)
+				}
+				if component.Metadata.Name == "" {
+					t.Errorf("component %d has empty metadata.name", i)
+				}
+
+				// Validate workloads have required fields
+				for j, workload := range component.Workloads {
+					if workload.Name == "" {
+						t.Errorf("component %d, workload %d has empty name", i, j)
+					}
+					if workload.Kind == "" {
+						t.Errorf("component %d, workload %d has empty kind", i, j)
+					}
+					if workload.DesiredImage == "" {
+						t.Errorf("component %d, workload %d has empty desiredImage", i, j)
+					}
+					// CurrentImage can be empty or "NOT_FOUND" if workload is not actually deployed
+				}
+
+				t.Logf("  Component [%d]: %s (%d workloads)", i+1, component.Metadata.Name, len(component.Workloads))
 			}
 		})
 	}
